% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{getDirSubList}
\alias{getDirSubList}
\title{Split a directory line into bits}
\usage{
getDirSubList(x)
}
\arguments{
\item{x}{A data frame with a single directory in it
(can have more than one line but just one valid dir)}
}
\value{
a list with \code{directory} and \code{note} items
}
\description{
Split a directory line into bits
}

