% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelLines.R
\name{labelLines}
\alias{labelLines}
\title{Annotate the lines in the make db}
\usage{
labelLines(x)
}
\arguments{
\item{x}{char(n) Character vector, probably from readLines}
}
\value{
a data frame with the original lines in \code{Lines} and various flags in other columns
}
\description{
Annotate the lines in the make db
}

