% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeParse.R
\name{makeParse}
\alias{makeParse}
\title{Parse the make database}
\usage{
makeParse(mkdb = NULL)
}
\arguments{
\item{mkdb}{Path to a text file containing the result of \code{make -p}. If left
out then it will attempt to call \code{make -p} from the current working directory.}
}
\value{
A list with each section from the make file
}
\description{
Takes the text output from \code{make -p} and parses it into structured text
that is easier to query.
}
\examples{
\dontrun{
  makeParse("mkdb.txt")
  makeParse()
}
}

