% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{parseDirectories}
\alias{parseDirectories}
\title{Parse the directory section}
\usage{
parseDirectories(dirdf)
}
\arguments{
\item{dirdf}{the directory section from the split data frame}
}
\value{
a list with directory name and note in each sub item
}
\description{
Parse the directory section
}

