% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{parseFiles}
\alias{parseFiles}
\title{Parse the files section}
\usage{
parseFiles(dbfiles)
}
\arguments{
\item{dbfiles}{A data frame containing the files section}
}
\value{
A list with an item per file
}
\description{
Parse the files section
}

