% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{parseOneFile}
\alias{parseOneFile}
\title{Parse a single file}
\usage{
parseOneFile(f)
}
\arguments{
\item{f}{A data frame containing all of the lines that are relevant to a single file}
}
\value{
A list with the file information
}
\description{
Usually called from an lapply. This is for pulling apart all the lines related
 to a single file and returning that in a list.
}

