% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{parseVar}
\alias{parseVar}
\title{Parse a variable record}
\usage{
parseVar(x)
}
\arguments{
\item{x}{A 2 row data frame with at least a Line column}
}
\value{
A list which the variable name, type and metadata all extracted
}
\description{
Parse a variable record
}

