% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{parseVariables}
\alias{parseVariables}
\title{Parse the variables section of the database}
\usage{
parseVariables(vardf)
}
\arguments{
\item{vardf}{The data frame corresponding to the variables section}
}
\value{
A list of lists, each containing a variable and its metadata
}
\description{
Things like environment variables and defaults. Keep as much metadata as possible
}

