% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sectionDropHeader}
\alias{sectionDropHeader}
\title{Strip the section name and following blank line}
\usage{
sectionDropHeader(dbsec)
}
\arguments{
\item{dbsec}{data frame for that section}
}
\value{
data.frame with section heading removed
}
\description{
Strip the section name and following blank line
}

