% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{sectionHeader}
\alias{sectionHeader}
\title{Get the name of a section}
\usage{
sectionHeader(dbsec)
}
\arguments{
\item{dbsec}{data frame for that section}
}
\value{
chr(1) Name of section
}
\description{
Get the name of a section
}

