% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{splitSections}
\alias{splitSections}
\title{Split the labelled data.frame into section}
\usage{
splitSections(dbdf)
}
\arguments{
\item{dbdf}{data.frame from \code{labelLines}}
}
\value{
a named list with each section as a data.frame
}
\description{
Split the labelled data.frame into section
}

